Fixing a bug in the byte array field codec

+++ server/src/main/java/com/jamierf/rsc/server/net/codec/field/ByteArrayFieldCodec.java
@@ -15,7 +15,7 @@ public class ByteArrayFieldCodec extends FieldCodec<byte[]> {
 
     @Override
     public void encode(byte[] value, ChannelBuffer buffer) {
-        buffer.writeByte(value.length);
+        buffer.writeShort(value.length);
         buffer.writeBytes(value);
     }
 }
